TW-315: Refactor i18n in BG. Refactor.
