+++ src/lib/temple/helpers.ts
@@ -1,12 +1,9 @@
-import { HttpResponseError } from '@taquito/http-utils';
 import { ManagerKeyResponse, RpcClient } from '@taquito/rpc';
 import { MichelCodecPacker } from '@taquito/taquito';
 import { validateAddress, ValidationResult } from '@taquito/utils';
 import BigNumber from 'bignumber.js';
 import memoize from 'micro-memoize';
 
-import { fetchMessage } from 'lib/i18n/for-bg';
-import { IntercomError } from 'lib/intercom/helpers';
 import { FastRpcClient } from 'lib/taquito-fast-rpc';
 
 export const loadFastRpcClient = memoize((rpc: string) => new FastRpcClient(rpc));
@@ -77,42 +74,4 @@ export function formatOpParamsBeforeSend(params: any) {
     return newParams;
   }
   return params;
-}
-
-export async function transformHttpResponseError(err: HttpResponseError) {
-  let parsedBody: any;
-  try {
-    parsedBody = JSON.parse(err.body);
-  } catch {
-    throw new Error(await fetchMessage('unknownErrorFromRPC', err.url));
-  }
-
-  try {
-    const firstTezError = parsedBody[0];
-
-    let message: string;
-
-    // Parse special error with Counter Already Used
-    if (typeof firstTezError.msg === 'string' && /Counter.*already used for contract/.test(firstTezError.msg)) {
-      message = await fetchMessage('counterErrorDescription');
-    } else {
-      const msgId = getTezErrLocaleMsgId(firstTezError?.id);
-      message = msgId ? await fetchMessage(msgId) : err.message;
-    }
-
-    return new IntercomError(message, parsedBody);
-  } catch {
-    throw err;
-  }
-}
-
-enum KNOWN_TEZ_ERRORS {
-  'implicit.empty_implicit_contract' = 'emptyImplicitContract',
-  'contract.balance_too_low' = 'balanceTooLow'
-}
-
-function getTezErrLocaleMsgId(tezErrId?: string) {
-  const idPostfixes = Object.keys(KNOWN_TEZ_ERRORS) as (keyof typeof KNOWN_TEZ_ERRORS)[];
-  const matchingPostfix = tezErrId && idPostfixes.find(idPostfix => tezErrId.endsWith(idPostfix));
-  return (matchingPostfix && KNOWN_TEZ_ERRORS[matchingPostfix]) || null;
 }
